+++ Panels/JVChatConsolePanel.m
@@ -463,9 +463,9 @@ static NSString *JVToolbarClearItemIdentifier = @"JVToolbarClearItem";
 	NSRect splitViewFrame = [splitView frame];
 	NSSize contentSize = [send minimumSizeForContent];
 	NSRect sendFrame = [[send enclosingScrollView] frame];
-	float dividerThickness = [splitView dividerThickness];
-	float maxContentHeight = ( NSHeight( splitViewFrame ) - dividerThickness - 75. );
-	float newContentHeight =  MIN( maxContentHeight, MAX( 22., contentSize.height + 8. ) );
+	CGFloat dividerThickness = [splitView dividerThickness];
+	CGFloat maxContentHeight = ( NSHeight( splitViewFrame ) - dividerThickness - 75. );
+	CGFloat newContentHeight =  MIN( maxContentHeight, MAX( 22., contentSize.height + 8. ) );
 
 	if( newContentHeight == NSHeight( sendFrame ) ) return;
 
@@ -616,7 +616,7 @@ static NSString *JVToolbarClearItemIdentifier = @"JVToolbarClearItem";
 }
 
 - (void) splitView:(NSSplitView *) sender resizeSubviewsWithOldSize:(NSSize) oldSize {
-	float dividerThickness = [sender dividerThickness];
+	CGFloat dividerThickness = [sender dividerThickness];
 	NSRect newFrame = [sender frame];
 
 	// Keep the size of the send box constant during window resizes
